lambda/python/requirements.txt
# 画像合成REST API Lambda関数の依存関係 - v2.4.0

# 画像処理ライブラリ
Pillow==10.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# HTTP通信
requests==2.31.0

# 日付・時刻処理
python-dateutil==2.8.2

# JSON処理（標準ライブラリだが明示的に記載）
# json - 標準ライブラリ

# UUID生成（標準ライブラリだが明示的に記載）
# uuid - 標準ライブラリ

# 並列処理（標準ライブラリだが明示的に記載）
# concurrent.futures - 標準ライブラリ

# ログ処理（標準ライブラリだが明示的に記載）
# logging - 標準ライブラリ

# 型ヒント（標準ライブラリだが明示的に記載）
# typing - 標準ライブラリ